delete command asks y/N if --force is not set

+++ digdag-cli/src/main/java/io/digdag/cli/client/Delete.java
@@ -1,6 +1,8 @@
 package io.digdag.cli.client;
 
 import java.io.PrintStream;
+import java.io.InputStreamReader;
+import java.io.BufferedReader;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -9,28 +11,22 @@ import java.util.HashMap;
 import java.io.File;
 
 import com.beust.jcommander.Parameter;
-import com.beust.jcommander.DynamicParameter;
-import com.google.inject.Injector;
-import com.google.inject.Scopes;
-import io.digdag.cli.Run;
-import io.digdag.cli.StdErr;
-import io.digdag.cli.StdOut;
 import io.digdag.cli.SystemExitException;
-import io.digdag.cli.YamlMapper;
+import io.digdag.cli.TimeUtil;
 import io.digdag.client.DigdagClient;
 import io.digdag.client.api.RestProject;
 import io.digdag.client.config.Config;
-import io.digdag.client.config.ConfigFactory;
 import io.digdag.core.DigdagEmbed;
 import io.digdag.core.Version;
-import io.digdag.core.config.ConfigLoaderManager;
 
-import static io.digdag.cli.Arguments.loadParams;
 import static io.digdag.cli.SystemExitException.systemExit;
 
 public class Delete
     extends ClientCommand
 {
+    @Parameter(names = {"--force"})
+    boolean force = false;
+
     public Delete(Version version, PrintStream out, PrintStream err)
     {
         super(version, out, err);
@@ -50,6 +46,7 @@ public class Delete
     {
         err.println("Usage: digdag delete <project>");
         err.println("  Options:");
+        err.println("        --force                      skip y/N prompt");
         showCommonOptions();
         return systemExit(error);
     }
@@ -60,6 +57,24 @@ public class Delete
         DigdagClient client = buildClient();
 
         RestProject proj = client.getProject(projName);
+
+        ln("Project:");
+        ln("  id: %d", proj.getId());
+        ln("  name: %s", proj.getName());
+        ln("  latest revision: %s", proj.getRevision());
+        ln("  created at: %s", TimeUtil.formatTime(proj.getCreatedAt()));
+        ln("  last updated at: %s", TimeUtil.formatTime(proj.getUpdatedAt()));
+
+        if (!force) {
+            err.print("Are you sure you want to delete this project? [y/N]: ");
+            String line;
+            try (BufferedReader stdin = new BufferedReader(new InputStreamReader(System.in))) {
+                line = stdin.readLine();
+            }
+            if (!line.trim().equalsIgnoreCase("y") && !line.trim().equalsIgnoreCase("yes")) {
+                throw systemExit("canceled.");
+            }
+        }
 
         client.deleteProject(proj.getId());
         err.println("Project '" + projName + "' is deleted.");
